renaming files so they match develop
