+++ src/app/cardsView/base-scrollable-cards-view.component.ts
